model/requirements.txt
inflection
pyyaml
